Move Span to new file
