+++ src/span.rs
@@ -1,0 +1,19 @@
+use std::fmt;
+
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub struct Span {
+    pub start: usize,
+    pub end: usize,
+}
+
+impl Span {
+    pub fn new(start: usize, end: usize) -> Span {
+        Span { start, end, }
+    }
+}
+
+impl fmt::Display for Span {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "start: {}, end: {}", self.start, self.end)
+    }
+}
