+++ src/token.rs
@@ -1,22 +1,6 @@
 use std::fmt;
 
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Span {
-    pub start: usize,
-    pub end: usize,
-}
-
-impl Span {
-    pub fn new(start: usize, end: usize) -> Span {
-        Span { start, end, }
-    }
-}
-
-impl fmt::Display for Span {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "start: {}, end: {}", self.start, self.end)
-    }
-}
+use crate::span::Span;
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub enum TokenType {
